backend/requirements.txt
# Django base
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Authentication
djangorestframework-simplejwt==5.3.1
djoser==2.2.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# API documentation
drf-spectacular==0.27.0

# Environment variables
python-dotenv==1.0.0

# Development & testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
Pillow==10.2.0
django-simple-history==3.4.0
django-storages==1.14.2